Added class Bullet

+++ src/Renderer/VertexBufferLayout.cpp
@@ -12,7 +12,7 @@ namespace RenderEngine
 	}
 	void VertexBufferLayout::addElementLayoutFloat(const unsigned int count, const bool normalized)
 	{
-		m_layoutElements.push_back({ count, GL_FLOAT, normalized, count * sizeof(GLfloat) });
+		m_layoutElements.push_back({ count, GL_FLOAT, normalized, count * static_cast<unsigned int>(sizeof(GLfloat)) });
 		m_stride += m_layoutElements.back().size;
 	}
 }
